refactor(domain): 장기판 초기화 메서드 리팩터링
